% Part of the R ica package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{ica-internal}

%% List of Internal Functions Called
\alias{sdiag}


\title{Internal functions for ica package}
\description{
  Internal functions for ica package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

